Formatting Bug
Sorry, I had changed the formatting a bit when running it locally. Specifically, I had removed the "date" field in the output file name, but it throws an error if the scenario includes the date. This commit brings back the "date" field, so it runs with the scenarios as they are formatted in this repository.